+++ src/decoder.rs
@@ -21,6 +21,8 @@ use crate::{
 
 #[derive(Debug)]
 pub struct AudioDecoder {
+    #[cfg(target_os = "linux")]
+    fdk_aac_lib: Option<shiguredo_fdk_aac::FdkAacLibrary>,
     engine_metric: crate::stats::StatsString,
     codec_metric: crate::stats::StatsString,
     total_audio_data_count_metric: crate::stats::StatsCounter,
@@ -36,12 +38,17 @@ enum DecoderRunOutput {
 }
 
 impl AudioDecoder {
-    pub fn new(mut compose_stats: crate::stats::Stats) -> crate::Result<Self> {
+    pub fn new(
+        #[cfg(target_os = "linux")] fdk_aac_lib: Option<shiguredo_fdk_aac::FdkAacLibrary>,
+        mut compose_stats: crate::stats::Stats,
+    ) -> crate::Result<Self> {
         let engine_metric = compose_stats.string("engine");
         let codec_metric = compose_stats.string("codec");
         let total_audio_data_count_metric = compose_stats.counter("total_audio_data_count");
         compose_stats.flag("error").set(false);
         Ok(Self {
+            #[cfg(target_os = "linux")]
+            fdk_aac_lib,
             engine_metric,
             codec_metric,
             total_audio_data_count_metric,
@@ -100,7 +107,13 @@ impl AudioDecoder {
 
         // 遅延初期化
         if self.inner.is_none() {
-            let inner = AudioDecoderInner::new(&frame)?;
+            let inner = AudioDecoderInner::new(
+                &frame,
+                #[cfg(target_os = "linux")]
+                self.fdk_aac_lib.take(),
+                #[cfg(not(target_os = "linux"))]
+                None,
+            )?;
             self.engine_metric.set(inner.engine_name().as_str());
             self.codec_metric.set(inner.codec_name().as_str());
             self.inner = Some(inner);
@@ -134,15 +147,12 @@ impl AudioDecoder {
         }
     }
 
-    pub fn get_engines(codec: CodecName) -> Vec<EngineName> {
+    pub fn get_engines(codec: CodecName, is_fdk_aac_available: bool) -> Vec<EngineName> {
         match codec {
             CodecName::Aac => {
-                // cfg によって mut が必要だったり不要だったりするので警告は抑制する
-                #[allow(unused_mut)]
                 let mut engines = Vec::new();
 
-                #[cfg(feature = "fdk-aac")]
-                {
+                if is_fdk_aac_available {
                     engines.push(EngineName::FdkAac);
                 }
                 #[cfg(target_os = "macos")]
@@ -163,27 +173,32 @@ enum AudioDecoderInner {
     Opus(OpusDecoder),
     #[cfg(target_os = "macos")]
     AudioToolbox(crate::decoder_audio_toolbox::AudioToolboxDecoder),
-    #[cfg(feature = "fdk-aac")]
+    #[cfg(target_os = "linux")]
     FdkAac(crate::decoder_fdk_aac::FdkAacDecoder),
 }
 
 impl AudioDecoderInner {
-    fn new(frame: &AudioFrame) -> crate::Result<Self> {
+    fn new(
+        frame: &AudioFrame,
+        #[cfg(target_os = "linux")] fdk_aac_lib: Option<shiguredo_fdk_aac::FdkAacLibrary>,
+        #[cfg(not(target_os = "linux"))] _fdk_aac_lib: Option<()>,
+    ) -> crate::Result<Self> {
         match frame.format {
             AudioFormat::Opus => OpusDecoder::new().map(Self::Opus),
             AudioFormat::Aac => {
-                #[cfg(feature = "fdk-aac")]
-                {
-                    crate::decoder_fdk_aac::FdkAacDecoder::new().map(Self::FdkAac)
+                #[cfg(target_os = "linux")]
+                if let Some(lib) = fdk_aac_lib {
+                    return crate::decoder_fdk_aac::FdkAacDecoder::new(lib).map(Self::FdkAac);
                 }
-                #[cfg(all(not(feature = "fdk-aac"), target_os = "macos"))]
+
+                #[cfg(target_os = "macos")]
                 {
                     crate::decoder_audio_toolbox::AudioToolboxDecoder::new().map(Self::AudioToolbox)
                 }
-                #[cfg(all(not(feature = "fdk-aac"), not(target_os = "macos")))]
+                #[cfg(not(any(target_os = "linux", target_os = "macos")))]
                 {
                     Err(crate::Error::new(
-                        "AAC decoding is only available on macOS or with fdk-aac feature enabled",
+                        "AAC decoding is not supported without --fdk-aac option or macOS",
                     ))
                 }
             }
@@ -199,7 +214,7 @@ impl AudioDecoderInner {
             Self::Opus(decoder) => decoder.decode(frame),
             #[cfg(target_os = "macos")]
             Self::AudioToolbox(decoder) => decoder.decode(frame),
-            #[cfg(feature = "fdk-aac")]
+            #[cfg(target_os = "linux")]
             Self::FdkAac(decoder) => decoder.decode(frame),
         }
     }
@@ -209,7 +224,7 @@ impl AudioDecoderInner {
             Self::Opus(_decoder) => Ok(None),
             #[cfg(target_os = "macos")]
             Self::AudioToolbox(decoder) => decoder.finish(),
-            #[cfg(feature = "fdk-aac")]
+            #[cfg(target_os = "linux")]
             Self::FdkAac(_decoder) => Ok(None),
         }
     }
@@ -219,7 +234,7 @@ impl AudioDecoderInner {
             Self::Opus(_) => EngineName::Opus,
             #[cfg(target_os = "macos")]
             Self::AudioToolbox(_) => EngineName::AudioToolbox,
-            #[cfg(feature = "fdk-aac")]
+            #[cfg(target_os = "linux")]
             Self::FdkAac(_) => EngineName::FdkAac,
         }
     }
@@ -229,7 +244,7 @@ impl AudioDecoderInner {
             Self::Opus(_) => CodecName::Opus,
             #[cfg(target_os = "macos")]
             Self::AudioToolbox(_) => CodecName::Aac,
-            #[cfg(feature = "fdk-aac")]
+            #[cfg(target_os = "linux")]
             Self::FdkAac(_) => CodecName::Aac,
         }
     }