+++ src/sora_recording_subcommand_compose.rs
@@ -29,6 +29,8 @@ struct Args {
     output_file_path: Option<PathBuf>,
     stats_file_path: Option<PathBuf>,
     openh264: Option<PathBuf>,
+    #[cfg(target_os = "linux")]
+    fdk_aac: Option<PathBuf>,
     no_progress_bar: bool,
     worker_threads: NonZeroUsize,
     root_dir: PathBuf,
@@ -62,6 +64,13 @@ impl Args {
                 .ty("PATH")
                 .env("HISUI_OPENH264_PATH")
                 .doc("OpenH264 の共有ライブラリのパスを指定します")
+                .take(raw_args)
+                .present_and_then(|a| a.value().parse())?,
+            #[cfg(target_os = "linux")]
+            fdk_aac: noargs::opt("fdk-aac")
+                .ty("PATH")
+                .env("HISUI_FDK_AAC_PATH")
+                .doc("FDK-AAC の共有ライブラリのパスを指定します")
                 .take(raw_args)
                 .present_and_then(|a| a.value().parse())?,
             no_progress_bar: noargs::flag("no-progress-bar")
@@ -124,6 +133,14 @@ fn run_internal(args: Args) -> crate::Result<()> {
         None
     };
 
+    // 必要に応じて FDK-AAC の共有ライブラリを読み込む
+    #[cfg(target_os = "linux")]
+    let fdk_aac_lib = args
+        .fdk_aac
+        .as_ref()
+        .map(shiguredo_fdk_aac::FdkAacLibrary::load)
+        .transpose()?;
+
     // 出力ファイルパスを決定
     let output_file_path = args
         .output_file_path
@@ -131,6 +148,8 @@ fn run_internal(args: Args) -> crate::Result<()> {
     let result = run_compose(
         layout,
         openh264_lib,
+        #[cfg(target_os = "linux")]
+        fdk_aac_lib,
         !args.no_progress_bar,
         args.worker_threads,
         output_file_path.clone(),
@@ -182,6 +201,7 @@ struct ComposePipelineSetup {
 fn run_compose(
     layout: Layout,
     openh264_lib: Option<Openh264Library>,
+    #[cfg(target_os = "linux")] fdk_aac_lib: Option<shiguredo_fdk_aac::FdkAacLibrary>,
     show_progress_bar: bool,
     worker_threads: NonZeroUsize,
     out_file_path: PathBuf,
@@ -196,6 +216,8 @@ fn run_compose(
     let result = runtime.block_on(run_compose_pipeline(
         layout,
         openh264_lib,
+        #[cfg(target_os = "linux")]
+        fdk_aac_lib,
         show_progress_bar,
         out_file_path,
     ))?;
@@ -230,6 +252,7 @@ fn run_compose(
 async fn run_compose_pipeline(
     layout: Layout,
     openh264_lib: Option<Openh264Library>,
+    #[cfg(target_os = "linux")] fdk_aac_lib: Option<shiguredo_fdk_aac::FdkAacLibrary>,
     show_progress_bar: bool,
     out_file_path: PathBuf,
 ) -> Result<ComposeResult> {
@@ -241,6 +264,8 @@ async fn run_compose_pipeline(
         &pipeline_handle,
         &layout,
         openh264_lib,
+        #[cfg(target_os = "linux")]
+        fdk_aac_lib,
         show_progress_bar,
         out_file_path,
     )
@@ -287,6 +312,7 @@ async fn setup_pipeline(
     pipeline_handle: &crate::MediaPipelineHandle,
     layout: &Layout,
     openh264_lib: Option<Openh264Library>,
+    #[cfg(target_os = "linux")] fdk_aac_lib: Option<shiguredo_fdk_aac::FdkAacLibrary>,
     show_progress_bar: bool,
     out_file_path: PathBuf,
 ) -> Result<ComposePipelineSetup> {
@@ -332,12 +358,18 @@ async fn setup_pipeline(
         let decoder_output_track_id = TrackId::new(decoder_processor_id.get());
         let reader_output_track_id_for_decoder = reader_output_track_id.clone();
         let decoder_output_track_id_for_decoder = decoder_output_track_id.clone();
+        #[cfg(target_os = "linux")]
+        let fdk_aac_lib_for_decoder = fdk_aac_lib.clone();
         spawn_processor_task(
             pipeline_handle,
             decoder_processor_id,
             decoder_metadata,
             move |handle| async move {
-                let decoder = AudioDecoder::new(handle.stats())?;
+                let decoder = AudioDecoder::new(
+                    #[cfg(target_os = "linux")]
+                    fdk_aac_lib_for_decoder,
+                    handle.stats(),
+                )?;
                 decoder
                     .run(
                         handle,
@@ -479,7 +511,13 @@ async fn setup_pipeline(
         audio_encoder_processor_id,
         audio_encoder_metadata,
         move |handle| async move {
-            let encoder = AudioEncoder::new(audio_codec, audio_bitrate, handle.stats())?;
+            let encoder = AudioEncoder::new(
+                audio_codec,
+                audio_bitrate,
+                #[cfg(target_os = "linux")]
+                fdk_aac_lib,
+                handle.stats(),
+            )?;
             encoder
                 .run(
                     handle,