+++ src/sora_recording_subcommand_vmaf.rs
@@ -32,6 +32,8 @@ struct Args {
     vmaf_output_file_path: Option<PathBuf>,
     openh264: Option<PathBuf>,
     #[expect(dead_code)]
+    fdk_aac: Option<PathBuf>,
+    #[expect(dead_code)]
     max_cpu_cores: Option<NonZeroUsize>,
     frame_count: usize,
     timeout: Option<Duration>,
@@ -71,6 +73,12 @@ impl Args {
                 .ty("PATH")
                 .env("HISUI_OPENH264_PATH")
                 .doc("OpenH264 の共有ライブラリのパスを指定します")
+                .take(raw_args)
+                .present_and_then(|a| a.value().parse())?,
+            fdk_aac: noargs::opt("fdk-aac")
+                .ty("PATH")
+                .env("HISUI_FDK_AAC_PATH")
+                .doc("FDK-AAC の共有ライブラリのパスを指定します")
                 .take(raw_args)
                 .present_and_then(|a| a.value().parse())?,
             max_cpu_cores: noargs::opt("max-cpu-cores")